+++ test/anthropic_test.exs
@@ -82,8 +82,11 @@ defmodule AnthropicTest do
   describe "request_next_message/1" do
     test "requires api_key" do
       Anthropic.Config.reset(api_key: nil)
+
       assert_raise(ArgumentError, fn ->
-        Anthropic.new(Anthropic.Config.reset(api_key: nil)) |> Anthropic.add_user_message("Good morning") |> Anthropic.request_next_message()
+        Anthropic.new(Anthropic.Config.reset(api_key: nil))
+        |> Anthropic.add_user_message("Good morning")
+        |> Anthropic.request_next_message()
       end)
     end
   end